requirements.txt
torch
transformers
gradio
Pillow
sentencepiece